Some progress made in test00

+++ 3phase/libuser.c
@@ -44,8 +44,8 @@ int Spawn(char *name, int (*func)(char *), char *arg, int stack_size,
     sysArg.number = SYS_SPAWN;
     sysArg.arg1 = (void *) func;
     sysArg.arg2 = arg;
-    sysArg.arg3 = (void *) stack_size;
-    sysArg.arg4 = (void *) priority;
+    sysArg.arg3 = ((void *) (long) stack_size);
+    sysArg.arg4 = ((void *) (long) priority);
     sysArg.arg5 = name;
 
     USLOSS_Syscall(&sysArg);
